requirements-oarec.txt
Flask
PyYAML